Refactor error handler file name
